+++ tests/bofire/data_models/features/test_engineered_feature.py
@@ -5,8 +5,10 @@ from bofire.data_models.features.api import (
     ContinuousDescriptorInput,
     ContinuousInput,
     ContinuousMolecularInput,
+    MolecularWeightedMeanFeature,
     MolecularWeightedSumFeature,
     SumFeature,
+    WeightedMeanFeature,
     WeightedSumFeature,
 )
 from bofire.data_models.molfeatures.api import MordredDescriptors
@@ -66,12 +68,9 @@ def test_weighted_sum_feature_validation():
         weighted_sum_feature.validate_features(inputs)
 
 
-def test_weighted_sum_feature_validation_with_normalization_flag():
-    weighted_sum_feature = WeightedSumFeature(
-        key="w_sum1",
-        features=["feat1", "feat2"],
-        descriptors=["desc1", "desc2"],
-        normalize_by_weight_sum=True,
+def test_weighted_mean_feature_validation():
+    weighted_mean_feature = WeightedMeanFeature(
+        key="w_mean1", features=["feat1", "feat2"], descriptors=["desc1", "desc2"]
     )
     inputs = Inputs(
         features=[
@@ -87,7 +86,7 @@ def test_weighted_sum_feature_validation_with_normalization_flag():
     with pytest.raises(
         ValueError, match="Feature 'feat2' is not a ContinuousDescriptorInput"
     ):
-        weighted_sum_feature.validate_features(inputs)
+        weighted_mean_feature.validate_features(inputs)
 
     inputs = Inputs(
         features=[
@@ -106,12 +105,38 @@ def test_weighted_sum_feature_validation_with_normalization_flag():
         ]
     )
     with pytest.raises(ValueError, match="Not all descriptors"):
-        weighted_sum_feature.validate_features(inputs)
+        weighted_mean_feature.validate_features(inputs)
 
 
 def test_molecular_weighted_sum_feature_validation():
     mol_feature = MolecularWeightedSumFeature(
         key="mw_sum1",
+        features=["m1", "m2"],
+        molfeatures=MordredDescriptors(descriptors=["NssCH2", "ATSC2d"]),
+    )
+    inputs = Inputs(
+        features=[
+            ContinuousMolecularInput(key="m1", bounds=(0, 1), molecule="C"),
+            ContinuousInput(key="m2", bounds=(0, 1)),
+        ]
+    )
+    with pytest.raises(
+        ValueError, match="Feature 'm2' is not a ContinuousMolecularInput"
+    ):
+        mol_feature.validate_features(inputs)
+
+    inputs = Inputs(
+        features=[
+            ContinuousMolecularInput(key="m1", bounds=(0, 1), molecule="C"),
+            ContinuousMolecularInput(key="m2", bounds=(0, 1), molecule="CC"),
+        ]
+    )
+    mol_feature.validate_features(inputs)
+
+
+def test_molecular_weighted_mean_feature_validation():
+    mol_feature = MolecularWeightedMeanFeature(
+        key="mw_mean1",
         features=["m1", "m2"],
         molfeatures=MordredDescriptors(descriptors=["NssCH2", "ATSC2d"]),
     )