+++ tests/bofire/data_models/specs/engineered_features.py
@@ -1,5 +1,5 @@
 from bofire.data_models.domain.api import EngineeredFeatures
-from bofire.data_models.features.api import MeanFeature, SumFeature, WeightedSumFeature
+from bofire.data_models.features.api import MeanFeature, SumFeature, WeightedMeanFeature
 from tests.bofire.data_models.specs.specs import Specs
 
 
@@ -11,11 +11,10 @@ specs.add_valid(
         "features": [
             SumFeature(key="sum1", features=["a", "b"]).model_dump(),
             MeanFeature(key="mean1", features=["a", "b"]).model_dump(),
-            WeightedSumFeature(
-                key="weighted_sum_normalized1",
+            WeightedMeanFeature(
+                key="weighted_mean1",
                 features=["a", "b"],
                 descriptors=["alpha", "beta"],
-                normalize_by_weight_sum=True,
             ).model_dump(),
         ],
     },